refactor(moonshot): split the trajectory verifier into focused modules, harden v2 chain validation
Review follow-ups on the v2 checkpointed chain format:

- Split the 856-line verifier per the repo's ~400-line module rule.
  verify-trajectory.mjs is now a thin CLI entry that dispatches on the
  chain version tag; the checks live in verify-common.mjs (failure shape,
  header/kernel pins), verify-v1.mjs (frozen per-step verifier),
  verify-v2.mjs (segment skeleton pass), verify-v2-segment.mjs (sampling,
  certificate shape, replay) and verify-endpoint.mjs (shared endpoint
  grounding). The v2 builder moves to chain-v2.mjs and the tamper battery
  splits the same way (tamper-test.mjs v1 + CLI, tamper-v2.mjs,
  tamper-common.mjs). Behavior-preserving: identical verdicts and
  timings, same seeded segment sample.
- Refuse nonpositive/fractional/NaN spot sample counts instead of
  sampling nothing and still reporting VERIFIED (invalid-spot-k; CLI
  exits 2, also for a bad --mode or --seed).
- Enforce the v2 segment-certificate contract before recomputing hashes:
  verifyCertificate accepts whatever a certificate asserts, so a hollowed
  certificate (reads/claims emptied, one honest end-root write) would
  pass on its own terms. checkSegmentCertificateShape now requires the
  entry and exit state node refs plus exactly one EmbodiedCarbon
  scalar-delta claim bound to the segment's own carbon endpoints, failing
  cleanly on malformed input rather than throwing.
- DESIGN.md: stamp the measurement table as a dated experiment record
  (date, machine, exact commands) rather than a live performance claim,
  and document the new module layout and validation rules.

Tamper battery: v2 19 cases pass (4 new certificate-shape cases, 4
spot-k guards), v1 6 cases still pass.
